check platform before accessing mrenclave file after docker build

+++ docker/build.sh
@@ -28,8 +28,10 @@ DOCKER_BUILDKIT=1 docker build \
     --build-arg SCITT_VERSION_OVERRIDE="$SCITT_VERSION_OVERRIDE" \
     .
 
-echo "mrenclave.txt"
-docker run --rm --entrypoint /bin/cat "$DOCKER_TAG" /usr/src/app/mrenclave.txt
+if [ "$PLATFORM" = "sgx" ]; then
+    echo "mrenclave.txt"
+    docker run --rm --entrypoint /bin/cat "$DOCKER_TAG" /usr/src/app/mrenclave.txt
+fi
 
 if [ -n "$SAVE_IMAGE_PATH" ]; then
     docker save "$DOCKER_TAG" -o "$SAVE_IMAGE_PATH"
